fix minor warnings
git-svn-id: svn+ssh://market.scs.stanford.edu/home/u1/asbestos/repo/jos64/trunk@1633 927ebb28-5305-0410-b3ca-c4b640df69b7

+++ pkg/uclibc/libc/sysdeps/jos64/fork.cc
@@ -127,7 +127,7 @@ do_fork()
     for (uint32_t i = 0; i < uas.nent; i++) {
 	if (uas.ents[i].flags == 0)
 	    continue;
-	if (uas.ents[i].segment.container == proc_ct)
+	if (uas.ents[i].segment.container == (uint64_t) proc_ct)
 	    continue;
 
 	// FDs are a special case
